requirements.txt
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
google-generativeai==0.3.2
python-dotenv==1.0.0
schedule==1.2.0
flask==3.0.0
flask-cors==4.0.0
pytz==2024.1
SpeechRecognition==3.10.0
gTTS==2.5.0
pydub==0.25.1
groq==0.11.0
sentence-transformers==2.2.2
pypdf==5.1.0
chromadb==0.4.22
